Crate trillium_client

trillium client is a http client that uses the same conn approach as trillium but which can be used independently for any http client application.

Connector

[trillium_client::Client] is built with a Connector. Each runtime crate (trillium_smol, trillium_tokio, trillium_async_std) offers a Connector implementation, which can optionally be combined with a tls crate such as trillium_rustls or trillium_native_tls.

See the documentation for Client and Conn for further usage examples.

Structs

Client

A client contains a Config and an optional connection pool and builds conns.

Conn

a client connection, representing both an outbound http request and a http response

UnexpectedStatusError

An unexpected http status code was received. Transform this back into the conn with From::from/Into::into.

Enums

Error

Concrete errors that occur within trillium’s http implementation

Traits

ClientLike

Trait for things that operate like a client. The only interface that’s required is build_conn.

Connector

Interface for runtime and tls adapters for the trillium client

Functions

client

constructs a new Client – alias for Client::new

Type Aliases

Result

this crate’s result type