Enum trillium_client::Error

#[non_exhaustive]
pub enum Error {

    Io(Error),
    UnexpectedUriFormat,
    HeaderMissing(&'static str),
    RequestPathMissing,
    Closed,
    Httparse(Error),
    TryFromIntError(TryFromIntError),
    PartialHead,
    MalformedHeader(Cow<'static, str>),
    UnsupportedVersion(u8),
    UnrecognizedMethod(String),
    MissingMethod,
    MissingStatusCode,
    UnrecognizedStatusCode(u16),
    MissingVersion,
    EncodingError(Utf8Error),
    UnexpectedHeader(&'static str),
    HeadersTooLong,
    ReceivedBodyTooLong(u64),

}

Concrete errors that occur within trilliumโs http implementation

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Io(Error)

std::io::Error

UnexpectedUriFormat

this error describes a malformed request with a path that does not start with / or http:// or https://

HeaderMissing(&'static str)

the relevant http protocol expected this header, but it was not provided

RequestPathMissing

this error describes a request that does not specify a path

Closed

connection was closed

Httparse(Error)

[httparse::Error]

TryFromIntError(TryFromIntError)

TryFromIntError

PartialHead

an incomplete http head

MalformedHeader(Cow<'static, str>)

we were unable to parse a header

UnsupportedVersion(u8)

async-h1 doesnโt speak this http version this error is deprecated

UnrecognizedMethod(String)

we were unable to parse this http method

MissingMethod

this request did not have a method

MissingStatusCode

this request did not have a status code

UnrecognizedStatusCode(u16)

we were unable to parse this http method

MissingVersion

this request did not have a version, but we expect one this error is deprecated

EncodingError(Utf8Error)

we expected utf8, but there was an encoding error

UnexpectedHeader(&'static str)

we received a header that does not make sense in context

HeadersTooLong

to mitigate against malicious http clients, we do not allow request headers beyond this length.

ReceivedBodyTooLong(u64)

to mitigate against malicious http clients, we do not read received bodies beyond this length to memory. If you need to receive longer bodies, use the Stream or AsyncRead implementation on ReceivedBody