Trait trillium_client::IntoUrl

pub trait IntoUrl {
    // Required method
    fn into_url(self, base: Option<&Url>) -> Result<Url>;
}

attempt to construct a url, with base if present

Required Methods

fn into_url(self, base: Option<&Url>) -> Result<Url>

attempt to construct a url, with base if present