Crate trillium_client

trillium client is a http client that uses the same conn approach as trillium but which can be used independently for any http client application.

Connector

[trillium_client::Client] is built with a Connector. Each runtime crate (trillium_smol, trillium_tokio, trillium_async_std) offers a Connector implementation, which can optionally be combined with a tls crate such as trillium_rustls or trillium_native_tls.

See the documentation for Client and Conn for further usage examples.

Structs

Client

A client contains a Config and an optional connection pool and builds conns.

Conn

a client connection, representing both an outbound http request and a http response

HeaderName

The name of a http header. This can be either a KnownHeaderName or a string representation of an unknown header.

HeaderValue

A HeaderValue represents the right hand side of a single name: value pair.

HeaderValues

A header value is a collection of one or more HeaderValue. It has been optimized for the “one HeaderValue” case, but can accomodate more than one value.

Headers

Trillium’s header map type

UnexpectedStatusError

An unexpected http status code was received. Transform this back into the conn with From::from/Into::into.

Url

A parsed URL record.

Enums

Error

Concrete errors that occur within trillium’s http implementation

KnownHeaderName

A short nonehaustive enum of headers that trillium can represent as a u8. Use a KnownHeaderName variant instead of a &’static str anywhere possible, as it allows trillium to skip parsing the header entirely.

Status

HTTP response status codes.

Version

The version of the HTTP protocol in use.

Constants

USER_AGENT

default http user-agent header

Traits

Connector

Interface for runtime and tls adapters for the trillium client

IntoUrl

attempt to construct a url, with base if present

Functions

client

constructs a new Client – alias for Client::new

Type Aliases

Result

this crate’s result type

Attribute Macros

async_trait